agrego funciones de iniciarViaje y cerrarViaje

+++ tpe-especial/microservicio-viajes/src/main/java/org/example/microservicioviajes/dto/ViajeDTO.java
@@ -14,24 +14,65 @@ public class ViajeDTO {
     private Long id;
     private LocalDateTime fechaInicio;
     private LocalDateTime fechaFin;
-    private BigDecimal kmRecorridos;
+    private Double kmRecorridos;
     private Long cuentaId;
     private Long monopatinId;
     private Long usuarioId;
 
     // Getters y setters explícitos (por si el analizador no procesa Lombok)
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
-    public LocalDateTime getFechaInicio() { return fechaInicio; }
-    public void setFechaInicio(LocalDateTime fechaInicio) { this.fechaInicio = fechaInicio; }
-    public LocalDateTime getFechaFin() { return fechaFin; }
-    public void setFechaFin(LocalDateTime fechaFin) { this.fechaFin = fechaFin; }
-    public BigDecimal getKmRecorridos() { return kmRecorridos; }
-    public void setKmRecorridos(BigDecimal kmRecorridos) { this.kmRecorridos = kmRecorridos; }
-    public Long getCuentaId() { return cuentaId; }
-    public void setCuentaId(Long cuentaId) { this.cuentaId = cuentaId; }
-    public Long getMonopatinId() { return monopatinId; }
-    public void setMonopatinId(Long monopatinId) { this.monopatinId = monopatinId; }
-    public Long getUsuarioId() { return usuarioId; }
-    public void setUsuarioId(Long usuarioId) { this.usuarioId = usuarioId; }
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public LocalDateTime getFechaInicio() {
+        return fechaInicio;
+    }
+
+    public void setFechaInicio(LocalDateTime fechaInicio) {
+        this.fechaInicio = fechaInicio;
+    }
+
+    public LocalDateTime getFechaFin() {
+        return fechaFin;
+    }
+
+    public void setFechaFin(LocalDateTime fechaFin) {
+        this.fechaFin = fechaFin;
+    }
+
+    public Double getKmRecorridos() {
+        return kmRecorridos;
+    }
+
+    public void setKmRecorridos(Double kmRecorridos) {
+        this.kmRecorridos = kmRecorridos;
+    }
+
+    public Long getCuentaId() {
+        return cuentaId;
+    }
+
+    public void setCuentaId(Long cuentaId) {
+        this.cuentaId = cuentaId;
+    }
+
+    public Long getMonopatinId() {
+        return monopatinId;
+    }
+
+    public void setMonopatinId(Long monopatinId) {
+        this.monopatinId = monopatinId;
+    }
+
+    public Long getUsuarioId() {
+        return usuarioId;
+    }
+
+    public void setUsuarioId(Long usuarioId) {
+        this.usuarioId = usuarioId;
+    }
 }
